Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartitionedStructures.M_elt_mat.Counter_elt_mat

PartitionedStructures.M_elt_mat.DenseEltMat

PartitionedStructures.M_elt_mat.Elt_mat

PartitionedStructures.M_elt_mat.LOEltMat

PartitionedStructures.M_elt_vec.Elt_vec

PartitionedStructures.M_internal_elt_vec.Internal_elt_vec

PartitionedStructures.M_internal_pv.Internal_pv

PartitionedStructures.M_part_mat.Part_LO_mat

PartitionedStructures.M_part_mat.Part_mat

PartitionedStructures.M_part_v.Part_v

PartitionedStructures.ModElemental_elom_bfgs.Elemental_elom_bfgs

PartitionedStructures.ModElemental_elom_sr1.Elemental_elom_sr1

PartitionedStructures.ModElemental_em.Elemental_em

PartitionedStructures.ModElemental_ev.Elemental_elt_vec

PartitionedStructures.ModElemental_plom.Elemental_plom

PartitionedStructures.ModElemental_plom_bfgs.Elemental_plom_bfgs

PartitionedStructures.ModElemental_plom_sr1.Elemental_plom_sr1

PartitionedStructures.ModElemental_pm.Elemental_pm

PartitionedStructures.ModElemental_pv.Elemental_pv

Base.permute!

Base.permute!

PartitionedStructures.Instances.create_epv_eplom

PartitionedStructures.Instances.create_epv_eplom_bfgs

PartitionedStructures.Instances.create_epv_eplom_sr1

PartitionedStructures.Instances.create_epv_epm

PartitionedStructures.Instances.create_epv_epm_rand

PartitionedStructures.Link.eplom_lbfgs_from_epv

PartitionedStructures.Link.eplom_lose_from_epv

PartitionedStructures.Link.eplom_lsr1_from_epv

PartitionedStructures.Link.epm_from_epv

PartitionedStructures.Link.epv_from_epm

PartitionedStructures.Link.mul_epm_epv

PartitionedStructures.Link.mul_epm_epv!

PartitionedStructures.Link.mul_epm_vector

PartitionedStructures.Link.mul_epm_vector

PartitionedStructures.Link.mul_epm_vector!

PartitionedStructures.Link.mul_epm_vector!

PartitionedStructures.Link.string_counters_iter

PartitionedStructures.Link.string_counters_total

PartitionedStructures.M_1_parallel.subproblem

PartitionedStructures.M_2_parallel.subproblem!

PartitionedStructures.M_3_parallel.subproblem3!

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.get_ee_struct

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_elt_mat.get_Bie

PartitionedStructures.M_elt_mat.get_cem

PartitionedStructures.M_elt_mat.get_counter_elt_mat

PartitionedStructures.M_elt_mat.iter_info

PartitionedStructures.M_elt_mat.total_info

PartitionedStructures.M_elt_mat.update_counter_elt_mat!

PartitionedStructures.M_frontale.frontale!

PartitionedStructures.M_internal_pv.rand_ipv

PartitionedStructures.M_okoubi_koko.okoubi

PartitionedStructures.M_okoubi_koko.okoubi!

PartitionedStructures.M_part_mat.get_ee_struct_Bie

PartitionedStructures.M_part_mat.get_eelom_set

PartitionedStructures.M_part_mat.get_eelom_set

PartitionedStructures.M_part_mat.get_eelom_set

PartitionedStructures.M_part_mat.get_eelom_set

PartitionedStructures.M_part_mat.get_eelom_set

PartitionedStructures.M_part_mat.get_eelom_set

PartitionedStructures.M_part_mat.get_permutation

PartitionedStructures.M_part_mat.get_spm

PartitionedStructures.M_part_mat.hard_reset_L!

PartitionedStructures.M_part_mat.hard_reset_spm!

PartitionedStructures.M_part_mat.reset_L!

PartitionedStructures.M_part_mat.reset_spm!

PartitionedStructures.M_part_mat.set_eelom_set!

PartitionedStructures.M_part_mat.set_permutation!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_v.add_v!

PartitionedStructures.M_part_v.build_v

PartitionedStructures.M_part_v.build_v!

PartitionedStructures.M_part_v.set_v!

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom_rand

PartitionedStructures.ModElemental_elom_bfgs.init_eelom_LBFGS

PartitionedStructures.ModElemental_elom_bfgs.reset_eelom_bfgs!

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom_rand

PartitionedStructures.ModElemental_elom_sr1.init_eelom_LSR1

PartitionedStructures.ModElemental_elom_sr1.reset_eelom_sr1!

PartitionedStructures.ModElemental_em.create_id_eem

PartitionedStructures.ModElemental_em.fixed_ones_eem

PartitionedStructures.ModElemental_em.identity_eem

PartitionedStructures.ModElemental_em.one_size_bloc

PartitionedStructures.ModElemental_em.ones_eem

PartitionedStructures.ModElemental_ev.create_eev

PartitionedStructures.ModElemental_ev.eev_from_sparse_vec

PartitionedStructures.ModElemental_ev.new_eev

PartitionedStructures.ModElemental_ev.ones_eev

PartitionedStructures.ModElemental_ev.sparse_vec_from_eev

PartitionedStructures.ModElemental_ev.specific_ones_eev

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom_rand

PartitionedStructures.ModElemental_plom.get_L

PartitionedStructures.ModElemental_plom.get_L

PartitionedStructures.ModElemental_plom.get_eelom_set_Bie

PartitionedStructures.ModElemental_plom.get_eelom_sub_set

PartitionedStructures.ModElemental_plom.identity_eplom_LOSE

PartitionedStructures.ModElemental_plom.set_L!

PartitionedStructures.ModElemental_plom.set_L_to_spm!

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom_rand

PartitionedStructures.ModElemental_plom_bfgs.get_L

PartitionedStructures.ModElemental_plom_bfgs.get_L

PartitionedStructures.ModElemental_plom_bfgs.get_eelom_set_Bie

PartitionedStructures.ModElemental_plom_bfgs.get_eelom_sub_set

PartitionedStructures.ModElemental_plom_bfgs.identity_eplom_LBFGS

PartitionedStructures.ModElemental_plom_bfgs.set_L!

PartitionedStructures.ModElemental_plom_bfgs.set_L_to_spm!

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom_rand

PartitionedStructures.ModElemental_plom_sr1.get_L

PartitionedStructures.ModElemental_plom_sr1.get_L

PartitionedStructures.ModElemental_plom_sr1.get_eelom_set_Bie

PartitionedStructures.ModElemental_plom_sr1.get_eelom_sub_set

PartitionedStructures.ModElemental_plom_sr1.identity_eplom_LSR1

PartitionedStructures.ModElemental_plom_sr1.set_L!

PartitionedStructures.ModElemental_plom_sr1.set_L_to_spm!

PartitionedStructures.ModElemental_pm.correlated_var

PartitionedStructures.ModElemental_pm.get_L

PartitionedStructures.ModElemental_pm.get_L

PartitionedStructures.ModElemental_pm.get_eem_set

PartitionedStructures.ModElemental_pm.get_eem_set

PartitionedStructures.ModElemental_pm.get_eem_set_Bie

PartitionedStructures.ModElemental_pm.get_eem_sub_set

PartitionedStructures.ModElemental_pm.identity_epm

PartitionedStructures.ModElemental_pm.identity_epm

PartitionedStructures.ModElemental_pm.identity_epm

PartitionedStructures.ModElemental_pm.n_i_SPS

PartitionedStructures.ModElemental_pm.n_i_sep

PartitionedStructures.ModElemental_pm.ones_epm

PartitionedStructures.ModElemental_pm.ones_epm_and_id

PartitionedStructures.ModElemental_pm.part_mat

PartitionedStructures.ModElemental_pm.set_L!

PartitionedStructures.ModElemental_pm.set_L_to_spm!

PartitionedStructures.ModElemental_pv.add_epv!

PartitionedStructures.ModElemental_pv.create_epv

PartitionedStructures.ModElemental_pv.epv_from_epv!

PartitionedStructures.ModElemental_pv.epv_from_v

PartitionedStructures.ModElemental_pv.epv_from_v!

PartitionedStructures.ModElemental_pv.minus_epv!

PartitionedStructures.ModElemental_pv.ones_kchained_epv

PartitionedStructures.ModElemental_pv.rand_epv

PartitionedStructures.ModElemental_pv.set_epv!

PartitionedStructures.PartMatInterface.update

PartitionedStructures.PartMatInterface.update!

PartitionedStructures.PartMatInterface.update!

PartitionedStructures.PartMatInterface.update!

PartitionedStructures.PartMatInterface.update!

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update!

PartitionedStructures.PartitionedLOQuasiNewton.PLSE_update

PartitionedStructures.PartitionedLOQuasiNewton.PLSE_update!

PartitionedStructures.PartitionedLOQuasiNewton.PLSR1_update

PartitionedStructures.PartitionedLOQuasiNewton.PLSR1_update!

PartitionedStructures.PartitionedLOQuasiNewton.Part_update

PartitionedStructures.PartitionedLOQuasiNewton.Part_update!

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update!

PartitionedStructures.PartitionedQuasiNewton.PSE_update

PartitionedStructures.PartitionedQuasiNewton.PSE_update!

PartitionedStructures.PartitionedQuasiNewton.PSR1_update

PartitionedStructures.PartitionedQuasiNewton.PSR1_update!

PartitionedStructures.Utils.BFGS

PartitionedStructures.Utils.SE

PartitionedStructures.Utils.SR1

​

PartitionedStructures.M_internal_elt_vec.Internal_elt_vec — Type

Internaleltvec{T} <: Elt_vec{T}

Type that represents an internal element vector.

PartitionedStructures.M_internal_pv.Internal_pv — Type

Internaleltvec{T} <: Elt_vec{T}

Type that represents an internal element vector.

PartitionedStructures.M_internal_pv.rand_ipv — Method

new_internal_pv(N,n;nᵢ,T)

Define an internal partitionned vector of N elemental nᵢ-sized vector simulating a n-sized T-vector.

PartitionedStructures.M_elt_vec.Elt_vec — Type

Abstract type representing element vectors.

PartitionedStructures.ModElemental_ev.Elemental_elt_vec — Type

Elementaleltvec{T} <: Elt_vec{T}

Type that represents an elemental element vector.

PartitionedStructures.ModElemental_ev.create_eev — Method

create_eev(vector_indices)

Create a random elemental element vector from vector_indices.

PartitionedStructures.ModElemental_ev.eev_from_sparse_vec — Method

eev_from_sparse_vec(sparse_vec)

eevfromsparsevec is an interface with SparseArrays.SparseVector. The indices and the values of the elemental element vector are define using findnz(sparsevec).

PartitionedStructures.ModElemental_ev.new_eev — Method

new_eev(nie; T, n)

Create an elemental element vector, with nie randoms values placed at indices within the range 1:n.

PartitionedStructures.ModElemental_ev.ones_eev — Method

ones_eev(nie; T, n)

Create an elemental element vector, with nie values at 1 placed at indices within the range 1:n.

PartitionedStructures.ModElemental_ev.sparse_vec_from_eev — Method

sparse_vec_from_eev(eev)

Create a sparse vector from the element element vector eev.

PartitionedStructures.ModElemental_ev.specific_ones_eev — Method

specific_ones_eev(nie, index; T, mul)

Create an elemental element vector, with nie randoms values multiplied by mul placed at indices in range index:index+nie, with .

PartitionedStructures.ModElemental_pv.Elemental_pv — Type

Elementalpv{T} <: Partv{T}

Type that represents an elemental partitioned vector.

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(epm)

initializecomponentlist! Build for each index i (∈ {1,...,n}) the list of the blocs using i.

PartitionedStructures.M_part_v.build_v! — Method

build_v!(pv)

Build from pv the vector v according to the information of each {evᵢ}ᵢ

PartitionedStructures.ModElemental_pv.add_epv! — Method

add_epv!(epv1,epv2)

Build in place of epv2 the addition of epv1 and epv2. Concretely each corresponding elemental vector will be add.

PartitionedStructures.ModElemental_pv.create_epv — Method

create_elemental_pv(elt_ev_set)

create an elemental partitioned vector from a collection (vector) of: sparse vector, elemental element vector or a vector of indices.

PartitionedStructures.ModElemental_pv.epv_from_epv! — Method

epv_from_epv!(evp1, epv2)

Set the elemental partitioned vector epv1 to epv2.

PartitionedStructures.ModElemental_pv.epv_from_v! — Method

epv_from_v!(epv, x)

Set the values of the element partitioned vector epv to x. Usefull to define Uᵢ x, ∀ x.

PartitionedStructures.ModElemental_pv.epv_from_v — Method

epv_from_v(x, epv)

Define a new elemental partitioned vector from x that have the same structure than epv. The value of each elemental element vector comes from the corresponding indices of x. Usefull to define Uᵢ x, ∀ x.

PartitionedStructures.ModElemental_pv.minus_epv! — Method

minus_epv!(epv)

Build in place the -epv, by inversing the value of each elemental element vector.

PartitionedStructures.ModElemental_pv.ones_kchained_epv — Method

ones_kchained_epv(N, k; T)

Construct a N-partitionned k-sized vector such as n = N+k.

PartitionedStructures.ModElemental_pv.rand_epv — Method

new_elemental_pv(N,n;nᵢ,T)

Define an elemental partitionned vector of N elemental nᵢ-sized vector simulating a n-sized T-vector.

PartitionedStructures.ModElemental_pv.set_epv! — Method

set_epv!(epv, vec_value_eev)

Set the values of the elemental element vectors of epv to the components of vec_value_eev.

PartitionedStructures.M_elt_mat.Counter_elt_mat — Type

Counter_elt_mat

Count for a element matrix the update performed on it, from its definition. total_update + total_reset + total_untouched == iter.

PartitionedStructures.M_elt_mat.DenseEltMat — Type

Abstract type representing a dense element matrix

PartitionedStructures.M_elt_mat.Elt_mat — Type

Abstract type representing an element matrix

PartitionedStructures.M_elt_mat.LOEltMat — Type

Abstract type representing a element linear operator

PartitionedStructures.M_elt_mat.get_Bie — Method

get_Bie(elt_mat)

Returns the element matrix elt_mat.Bie.

PartitionedStructures.M_elt_mat.get_cem — Method

get_cem(elt_mat)

Returns the Counter_elt_mat of the elemental element matrix elt_mat.

PartitionedStructures.M_elt_mat.get_counter_elt_mat — Method

get_counter_elt_mat(elt_mat)

Returns the Counter_elt_mat of the elemental element matrix elt_mat.

PartitionedStructures.M_elt_mat.iter_info — Method

iter_info(cem)

Returns the information about the last quasi-Newton update applied onto the element associated to the counter cem.

PartitionedStructures.M_elt_mat.total_info — Method

iter_info(cem)

Returns the informations about all the quasi-Newton updates applied onto the element associated to the counter cem.

PartitionedStructures.M_elt_mat.update_counter_elt_mat! — Method

update_counter_elt_mat!(cem, qn)

Updates the cem counter given the index qn from the quasi-Newton update.

PartitionedStructures.M_part_mat.Part_LO_mat — Type

Abstract type representing partitioned matrix using linear operators

PartitionedStructures.M_part_mat.Part_mat — Type

Abstract type representing partitioned matrix

PartitionedStructures.M_part_mat.get_ee_struct_Bie — Method

get_ee_struct_Bie(pm, i)

Returns the i-th elemental element matrix of the partitioned matrix pm.

PartitionedStructures.M_part_mat.get_permutation — Method

get_permutation(pm)

Gets the current permutation of the partitioned matrix pm.

PartitionedStructures.M_part_mat.get_spm — Method

get_spm(pm)

get_spm(pm, i, j)

Get either the sparse matrix associated to the partitioned matrix pm or pm[i,j].

PartitionedStructures.M_part_mat.hard_reset_L! — Method

hard_reset_L!(pm)

Reset the sparse matrix pm.L.

PartitionedStructures.M_part_mat.hard_reset_spm! — Method

hard_reset_spm!(pm)

Reset the sparse matrix pm.spm.

PartitionedStructures.M_part_mat.reset_L! — Method

reset_L!(pm)

Set the elements of sparse matrix pm.L to 0.

PartitionedStructures.M_part_mat.reset_spm! — Method

reset_spm!(pm)

Set the elements of sparse matrix pm.spm to 0.

PartitionedStructures.M_part_mat.set_permutation! — Method

get_permutation(pm, perm)

Set the permutation of the partitioned matrix pm to perm.

PartitionedStructures.M_part_v.Part_v — Type

Abstract type representing partitioned vectors.

PartitionedStructures.M_part_v.add_v! — Method

add_v!(pv, i, value)

add_v!(pv, indices, values)

Add value (resp values) to the vector associated to the partitioned vector pv.v at the indice i (resp indices).

PartitionedStructures.M_part_v.build_v — Method

build_v(pv)

Build the vector v from the partitionned vector pv. Call specialised method depending the type of the element vector inside pv For now if there is mix of elemental and internal element vectors it must be previously transform as internal partitioned vector.

PartitionedStructures.M_part_v.set_v! — Method

set_v!(pv, vec)

set_v!(pv, index, value)

Set to value the vector associated to the partitioned vector pv.v at the indice index. Set to vec the vector associated to the partitioned vector pv.v.

PartitionedStructures.ModElemental_pm.Elemental_pm — Type

Elementalpm{T} <: Partmat{T}

Type that represents an elemental partitioned quasi-Newton linear operator, each Bᵢ may use a BFGS or SR1 update.

Base.permute! — Method

permute!(epm, p)

Apply the permutation p to the elemental partitionned matrix epm. The permutation is applied to each elemental element matrice eem via indices. The current epm permutation is stored in epm.permutation.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(epm)

Returns the vector of every elemental element matrix epm.eem_set.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(epm, i)

Returns the i-th elemental element matrix epm.eem_set[i].

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(epm)

Build for each index i (∈ {1, ..., n}) a list of the elements using the i-th variable.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(epm)

Build the sparse matrix of eplom in eplom.spm from the blocs eplom.eelom_set. The sparse matrix is build according to the indices of each elemental element linear operator.

PartitionedStructures.ModElemental_pm.correlated_var — Method

correlated_var(epm, i)

Gets the variables that appears in the same elements than the i-th variable.

PartitionedStructures.ModElemental_pm.get_L — Method

get_L(epm)

Returns the sparse matrix epm.L, whose aim to store a cholesky factor. By default epm.L is not instantiate.

PartitionedStructures.ModElemental_pm.get_L — Method

get_L(epm, i, j)

Returns the value epm.L[i,j], from the sparse matrix epm.L.

PartitionedStructures.ModElemental_pm.get_eem_set — Method

get_eem_set(epm)

Returns the vector of every elemental element matrix epm.eem_set.

PartitionedStructures.ModElemental_pm.get_eem_set — Method

get_eem_set(epm :: Elemental_pm_bfgs{T}, i :: Int)

Returns the i-th elemental element matrix epm.eem_set[i].

PartitionedStructures.ModElemental_pm.get_eem_set_Bie — Method

get_eem_set_Bie(epm, i)

Get the linear operator of the i-th elemental element linear operator of epm.

PartitionedStructures.ModElemental_pm.get_eem_sub_set — Method

get_eem_set_Bie(epm, indices)

Returns a subset of the elemental element matrix composing epm. indices selects the differents elemental element matrix needed.

PartitionedStructures.ModElemental_pm.identity_epm — Method

identity_epm(N, n; T=T, nie)

Create a partitionned matrix of type T of N identity elemental element matrices. Each elemental element matrix is of size nie with randoms positions.

PartitionedStructures.ModElemental_pm.identity_epm — Method

identity_epm(element_variables, N, n; T=T)

Create a partitionned matrix of type T of N identity elemental element matrices. The elemental variables are based from the indices informed in element_variables.

PartitionedStructures.ModElemental_pm.identity_epm — Method

identity_epm(element_variables; N, n, T=T)

Create a partitionned matrix of type T of N identity elemental element matrices. N and n are extrapolate from element_variables. The elemental variables are based from the indices informed in element_variables.

PartitionedStructures.ModElemental_pm.n_i_SPS — Method

n_i_SPS(n; T, nie, overlapping, mul)

Defines an elemental partitioned matrix of size n. The partitioned matrix is composed by N ≈ (n/nie)*2 elemental element matrices, of size nie, overlapping onto the diagonal. The diagonal terms of each elemental element matrix are of value mul.

PartitionedStructures.ModElemental_pm.n_i_sep — Method

n_i_sep(n; T=T, nie=nie, mul=mul)

Define a partitioned nie bloc separable matrix. Each elemental element matrix is composed of 1 except the diagonal terms which are of value mul.

PartitionedStructures.ModElemental_pm.ones_epm — Method

ones_epm(N, n; T=T, nie=nie)

Create a partitionned matrix of type T of N elemental element matrices ones(nie, nie) whose positions are random. The partitionned matrix created may be singular.

PartitionedStructures.ModElemental_pm.ones_epm_and_id — Method

ones_epm_and_id(N, n; T=T, nie=nie)

Create a partitionned matrix of type T with N+n elemental element matrices. Each elemental element matrice is ones(nie, nie) with randoms positions in the range 1:n. The remaining n elemental element matrices are of size 1, with value [1], they are placed in the diagonal terms. This way, the partitionned matrix is generally not singular.

PartitionedStructures.ModElemental_pm.part_mat — Method

part_mat(;n=n, T=T, nie=nie, overlapping=overlapping, mul=mul)

Define a elemental partitioned matrix formed by N (deduced from n and nie) elemental element matrices of size nie. Each elemental element matrix overlaps the previous one and the next one by overlapping.

PartitionedStructures.ModElemental_pm.set_L! — Method

set_L!(epm, i, j, value)

Sets the value of epm.L[i,j] = value.

PartitionedStructures.ModElemental_pm.set_L_to_spm! — Method

set_L_to_spm!(epm, i, j, value)

Sets the sparse matrix epm.L to the sparse matrix epm.spm.

PartitionedStructures.ModElemental_em.Elemental_em — Type

Elemental_em{T} <: DenseEltMat{T}

Type that represents an elemental element matrix

Base.permute! — Method

permute!(eem, p)

Set the indices of the element variables of eem to p. Must be use with caution.

PartitionedStructures.ModElemental_em.create_id_eem — Method

create_id_eem(elt_var; T=T)

Creates a nie identity elemental element matrix of type T at the indices elt_var.

PartitionedStructures.ModElemental_em.fixed_ones_eem — Method

fixed_ones_eem(index, nie; type=T, mul=mul)

Creates a nie elemental element matrix of type T at indices index:index+nie-1. All the components of the element matrix are set to 1 except the diagonal terms that are set to mul. This method is used to define diagonal dominant element matrix.

PartitionedStructures.ModElemental_em.identity_eem — Method

identity_eem(nie; T=T, n=n)

Creates a nie identity elemental element matrix of type T from nie random indices in the range 1:n.

PartitionedStructures.ModElemental_em.one_size_bloc — Method

one_size_bloc(index)

Defines an elemental element matrix of type T of size one at index.

PartitionedStructures.ModElemental_em.ones_eem — Method

ones_eem(nie; T=T, n=n)

Creates a nie ones elemental element matrix of type T from nie random indices in the range 1:n.

PartitionedStructures.M_okoubi_koko.okoubi! — Method

okoubi!(epm, epv, epv_tmp, res)

Solve the partitioned linear problem using the Okoubi and Koko method described in the paper available at https://doi.org/10.1007/s11590-016-1020-x. This method avoid the allocation of epv_tmp and res.

PartitionedStructures.M_okoubi_koko.okoubi — Method

okoubi(epm, epv)

Solve the partitioned linear problem from epm and epv using the Okoubi and Koko method described in the paper available at https://doi.org/10.1007/s11590-016-1020-x.

PartitionedStructures.M_frontale.frontale! — Method

frontale!(epm)

Produce the Cholesky factorization of the elemental partitioned matrix epm using a frontal method. The sparse factor L is stored in epm.L.

PartitionedStructures.M_1_parallel.subproblem — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.M_2_parallel.subproblem! — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.M_3_parallel.subproblem3! — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.Utils.BFGS — Method

BFGS(s, y, B)

Perform the BFGS update over the matrix B by using the vector s and y.

PartitionedStructures.Utils.SE — Method

SE(s, y, B)

Perform either the BFGS update or SR1 over the matrix B by using the vector s and y.

PartitionedStructures.Utils.SR1 — Method

SR1(s, y, B)

Perform the BFGS update over the matrix B by using the vector s and y.

PartitionedStructures.Link.eplom_lbfgs_from_epv — Method

eplom = eplom_lbfgs_from_epv(epv)

Create an elemental partitioned limited-memory quasi-Newton operator PLBFGS eplom with the same partitioned structure than epv. Each element linear operator of eplom is set to a LBFGSOperator of suitable size.

PartitionedStructures.Link.eplom_lose_from_epv — Method

eplom = eplom_lose_from_epv(epv)

Create an elemental partitioned limited-memory quasi-Newton operator PLSR1 eplom with the same partitioned structure than epv. Each element linear operator of eplom is set to a LBFGSOperator of suitable size, but it may change to LSR1Operator later on.

PartitionedStructures.Link.eplom_lsr1_from_epv — Method

eplom = eplom_lsr1_from_epv(epv)

Create an elemental partitioned limited-memory quasi-Newton operator PLSR1 eplom with the same partitioned structure than epv. Each element linear operator of eplom is set to a LBFGSOperator of suitable size.

PartitionedStructures.Link.epm_from_epv — Method

epm = epm_from_epv(epv)

Create an elemental partitioned quasi-Newton operator epm with the same partitioned structure than epv. Each element matrix of epm is set with an identity matrix of suitable size.

PartitionedStructures.Link.epv_from_epm — Method

epv = epv_from_epm(epm)

Create an elemental partitioned vector epv with the same partitioned structure than epm. Each element vector of epv is set to a random vector of suitable size.

PartitionedStructures.Link.mul_epm_epv! — Method

mul_epm_epv!(epv_res, epm, epv)

Compute the product between the elemental partitioned matrix epm and the elemental partitioned vector epv. The result of each element matrix vector product is stored in the elemental partitioned vector epv_res.

PartitionedStructures.Link.mul_epm_epv — Method

mul_epm_epv(epm, epv)

Compute the product between the elemental partitioned matrix epm and the elemental partitioned vector epv.

PartitionedStructures.Link.mul_epm_vector! — Method

mul_epm_vector!(res, epm, epv, x)

Compute the product between the elemental partitioned matrix epm and the vector x using temporary the elemental partitioned vector epv. The result is stored in the vector res.

PartitionedStructures.Link.mul_epm_vector! — Method

mul_epm_vector!(res, epm, x)

It computes the product between the elemental partitioned matrix epm and the vector x. The result is stored in the vector res.

PartitionedStructures.Link.mul_epm_vector — Method

result = mul_epm_vector!(epm, epv, x)

It computes the product between the elemental partitioned matrix epm and the vector x using temporary the elemental partitioned vector epv.

PartitionedStructures.Link.mul_epm_vector — Method

result = mul_epm_vector(epm, x)

It computes the product between the elemental partitioned matrix epm <: Part_mat and the vector x. The method return the vector result.

PartitionedStructures.Link.string_counters_iter — Method

string_counters_iter(pm)

Produces a string that summarize the partitioned update applied onto pm at the last iterate.

PartitionedStructures.Link.string_counters_total — Method

string_counters_total(pm)

Produces a string that summarize the partitioned update applied onto pm from its allocation at the start of an algorithm.

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update! — Method

PBFGS_update!(epm_B, epv_y, epv_s)

Performs the partitioned BFGS update onto the partitioned matrix epm_B, given the element steps epv_s and the element gradient differences epv_y.

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update — Method

B = PBFGS_update(epm_B, epv_y, s)

Performs the partitioned BFGS update onto the partitioned matrix epm_B, given the step s and the element gradient difference epv_y

PartitionedStructures.PartitionedQuasiNewton.PSE_update! — Method

PSE_update!(epm_B, epv_y, epv_s)

Performs the partitioned update PSE of the elemental partitioned matrix epm_B, given the element steps epv_s and the element gradient differences epv_y.

PartitionedStructures.PartitionedQuasiNewton.PSE_update — Method

B = PSE_update(epm_B, epv_y, s)

Performs the partitioned update PSE of the elemental partitioned matrix epm_B, given the step s and the element gradient difference epv_y

PartitionedStructures.PartitionedQuasiNewton.PSR1_update! — Method

PSR1_update!(epm_B, epv_y, s)

Performs the partitioned-SR1 update of the elemental partitioned matrix epm_B, given the element steps epv_s and the element gradient differences epv_y.

PartitionedStructures.PartitionedQuasiNewton.PSR1_update — Method

B = PSR1_update(epm_B, epv_y, s)

Performs the partitioned SR1 update of the elemental partitioned matrix epm_B, given the step s and the element gradient difference epv_y

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update! — Method

PLBFGS_update!(eplom_B, s, epv_y)

Performs the partitioned-LBFGS update onto the limited-memory partitioned operator eplom_B, given the element steps epv_s and the element gradient differences epv_y.

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update — Method

B = PLBFGS_update(eplom_B, s, epv_y)

Performs the partitioned-LBFGS update onto the limited-memory partitioned operator eplom_B, given the step s and the element gradient differences epv_y.

PartitionedStructures.PartitionedLOQuasiNewton.PLSE_update! — Method

PLSE_update!(eplom_B, epv_y, epv_s)

Performs the partitionned update PLSE of eplom_B, which is made of LBFGS or LSR1 elemental element matrices, given the element steps epv_s and the element gradient differences epv_y. The update try to apply a LBFGS update to every Bᵢ, but if the curvature condition yᵢᵀs > 0 is not satisfied it applies LSR1 in place of LBFGS.

PartitionedStructures.PartitionedLOQuasiNewton.PLSE_update — Method

B = PLSE_update(eplom_B, epv_y, s)

Performs the partitionned update PLSE of eplom_B, which is made of LBFGS or LSR1 elemental element matrices, given the step s and the element gradient differences epv_y. The update try to apply a LBFGS update to every Bᵢ, but if the curvature condition yᵢᵀs > 0 is not satisfied it applies LSR1 in place of LBFGS.

PartitionedStructures.PartitionedLOQuasiNewton.PLSR1_update! — Method

PLSR1_update!(eplom_B, s, epv_y)

Performs the partitioned-SR1 update onto the limited-memory partitioned operator eplom_B, given the element steps epv_s and the element gradient differences epv_y.

PartitionedStructures.PartitionedLOQuasiNewton.PLSR1_update — Method

B = PLSR1_update(eplom_B, s, epv_y)

Performs the partitioned-SR1 update onto the limited-memory partitioned operator eplom_B, given the step s and the element gradient differences epv_y.

PartitionedStructures.PartitionedLOQuasiNewton.Part_update! — Method

Part_update!(eplom_B, epv_y, s)

Performs a quasi-Newton partitionned update of eplom_B, given the element steps epv_s and the element gradient differences epv_y. Each element linear operator from eplom_B is either a LBFGSOperator or LSR1Operator. The update performs on each element the quasi-Newton to the linear operator associated.

PartitionedStructures.PartitionedLOQuasiNewton.Part_update — Method

B = Part_update(eplom_B, epv_y, s)

Performs a quasi-Newton partitionned update of eplom_B, given the step s and the element gradient differences epv_y. Each element linear operator from eplom_B is either a LBFGSOperator or LSR1Operator. The update performs on each element the quasi-Newton to the linear operator associated.

PartitionedStructures.ModElemental_elom_bfgs.Elemental_elom_bfgs — Type

Elemental_elom_bfgs{T} <: LOEltMat{T}

Type that represents an elemental element linear operator LBFGS

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom — Method

LBFGS_eelom(nie; T=T, index=index)

Creates an Elemental_elom_bfgs of type T of size nie, the indices are all the values in the range index:index+nie-1.

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom_rand — Method

LBFGS_eelom_rand(nie; T=T, n=n)

Creates an Elemental_elom_bfgs of type T with nie random indices within the range 1:n.

PartitionedStructures.ModElemental_elom_bfgs.init_eelom_LBFGS — Method

init_eelom_LBFGS(elt_var; T=T)

Defines an Elemental_elom_bfgs of type T based from the vector of indices elt_var.

PartitionedStructures.ModElemental_elom_bfgs.reset_eelom_bfgs! — Method

reset_eelom_bfgs!(eelom)

Resets the LBFGS linear operator of the elemental element linear operator matrix eelom.

PartitionedStructures.ModElemental_plom_bfgs.Elemental_plom_bfgs — Type

Elemental_plom_bfgs{T} <: Part_LO_mat{T}

Type that represents an elemental limited-memory partitioned quasi-Newton operator PLBFGS.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(eplom)

Returns the vector of every elemental element linear operator eplom.eelom_set.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(eplom, i)

Returns the i-th elemental element linear operator eplom.eelom_set[i].

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

Build for each index i (∈ {1, ..., n}) a list of the elements using the i-th variable.

PartitionedStructures.M_part_mat.get_eelom_set — Method

get_eelom_set(eplom)

Returns the vector of every elemental element linear operator eplom.eelom_set.

PartitionedStructures.M_part_mat.get_eelom_set — Method

get_eelom_set(eplom :: Elemental_plom_bfgs{T}, i :: Int)

Returns the i-th elemental element linear operator eplom.eelom_set[i].

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix of eplom in eplom.spm from the blocs eplom.eelom_set. The sparse matrix is build according to the indices of each elemental element linear operator.

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom — Method

PLBFGS_eplom(;n, type, nie, overlapping)

Creates an elemental partitionned limited-memory operator PLBFGS of N (deduced from n and nie) elemental element linear operators. Each element overlaps the coordinates of the next element by overlapping components.

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom_rand — Method

PLBFGS_eplom_rand(N, n; type, nie)

Create an elemental partitionned limited-memory operator PLBFGS of N elemental element linear operators. The size of each element is nie, whose positions are random in the range 1:n.

PartitionedStructures.ModElemental_plom_bfgs.get_L — Method

get_L(eplom)

Returns the sparse matrix eplom.L, whose aim to store a cholesky factor. By default eplom.L is not instantiate.

PartitionedStructures.ModElemental_plom_bfgs.get_L — Method

get_L(eplom, i, j)

Returns the value eplom.L[i,j], from the sparse matrix eplom.L.

PartitionedStructures.ModElemental_plom_bfgs.get_eelom_set_Bie — Method

get_eelom_set_Bie(eplom, i)

Get the linear operator of the i-th elemental element linear operator of eplom.

PartitionedStructures.ModElemental_plom_bfgs.get_eelom_sub_set — Method

get_eelom_set_Bie(eplom, indices)

Returns a subset of the elemental element linear operators composing eplom. indices selects the differents elemental element linear operators needed.

PartitionedStructures.ModElemental_plom_bfgs.identity_eplom_LBFGS — Method

identity_eplom_LBFGS(element_variables, N, n; T=T)

Creates an elemental partitionned limited-memory operator PLBFGS of N elemental element linear operators. The positions are given by the vector of the element variables element_variables.

PartitionedStructures.ModElemental_plom_bfgs.set_L! — Method

set_L!(eplom, i, j, value)

Sets the value of eplom.L[i,j] = value.

PartitionedStructures.ModElemental_plom_bfgs.set_L_to_spm! — Method

set_L_to_spm!(eplom, i, j, value)

Sets the sparse matrix eplom.L to the sparse matrix eplom.spm.

PartitionedStructures.ModElemental_elom_sr1.Elemental_elom_sr1 — Type

Elemental_elom_sr1{T} <: LOEltMat{T}

Type that represents an elemental element linear operator LSR1

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom — Method

LSR1_eelom(nie, T=T, index=index)

Creates an Elemental_elom_sr1 of type T of size nie, the indices are all the values in the range index:index+nie-1.

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom_rand — Method

LSR1_eelom_rand(nie, T=T, n=n)

Creates an Elemental_elom_sr1 of type T with nie random indices within the range 1:n.

PartitionedStructures.ModElemental_elom_sr1.init_eelom_LSR1 — Method

init_eelom_LSR1(elt_var; T=T)

Defines an Elemental_elom_sr1 of type T based from the vector of indices elt_var.

PartitionedStructures.ModElemental_elom_sr1.reset_eelom_sr1! — Method

index_eelom_sr1!(eelom)

Resets the LSR1 linear operator of the elemental element linear operator matrix.

PartitionedStructures.ModElemental_plom_sr1.Elemental_plom_sr1 — Type

Elemental_plom_sr1{T} <: Part_LO_mat{T}

Type that represents an elemental limited-memory partitioned quasi-Newton operator PLSR1.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(eplom)

Returns the vector of every elemental element linear operator eplom.eelom_set.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(eplom, i)

Returns the i-th elemental element linear operator eplom.eelom_set[i].

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

Build for each index i (∈ {1, ..., n}) a list of the elements using the i-th variable.

PartitionedStructures.M_part_mat.get_eelom_set — Method

get_eelom_set(eplom)

Returns the vector of every elemental element linear operator eplom.eelom_set.

PartitionedStructures.M_part_mat.get_eelom_set — Method

get_eelom_set(eplom :: Elemental_plom_bfgs{T}, i :: Int)

Returns the i-th elemental element linear operator eplom.eelom_set[i].

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix of eplom in eplom.spm from the blocs eplom.eelom_set. The sparse matrix is build according to the indices of each elemental element linear operator.

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom — Method

PLSR1_eplom(N, n; type, nie)

Creates an elemental partitionned limited-memory operator PLSR1 of N (deduced from n and nie) elemental element linear operators. Each element overlaps the coordinates of the next element by overlapping components.

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom_rand — Method

PLSR1_eplom_rand(N, n; type, nie)

Create an elemental partitionned limited-memory operator PLSR1 of N elemental element linear operators. The size of each element is nie, whose positions are random in the range 1:n.

PartitionedStructures.ModElemental_plom_sr1.get_L — Method

get_L(eplom)

Returns the sparse matrix eplom.L, whose aim to store a cholesky factor. By default eplom.L is not instantiate.

PartitionedStructures.ModElemental_plom_sr1.get_L — Method

get_L(eplom, i, j)

Returns the value eplom.L[i,j], from the sparse matrix eplom.L.

PartitionedStructures.ModElemental_plom_sr1.get_eelom_set_Bie — Method

get_eelom_set_Bie(eplom, i)

Get the linear operator of the i-th elemental element linear operator of eplom.

PartitionedStructures.ModElemental_plom_sr1.get_eelom_sub_set — Method

get_eelom_set_Bie(eplom, indices)

Returns a subset of the elemental element linear operators composing eplom. indices selects the differents elemental element linear operators needed.

PartitionedStructures.ModElemental_plom_sr1.identity_eplom_LSR1 — Method

identity_eplom_LSR1(vec_indices, N, n; T=T)

Creates an elemental partitionned limited-memory operator PLSR1 of N elemental element linear operators. The positions are given by the vector of the element variables element_variables.

PartitionedStructures.ModElemental_plom_sr1.set_L! — Method

set_L!(eplom, i, j, value)

Sets the value of eplom.L[i,j] = value.

PartitionedStructures.ModElemental_plom_sr1.set_L_to_spm! — Method

set_L_to_spm!(eplom, i, j, value)

Sets the sparse matrix eplom.L to the sparse matrix eplom.spm.

PartitionedStructures.ModElemental_plom.Elemental_plom — Type

Elementalplomsr1{T} <: PartLOmat{T}

Type that represents an elemental limited-memory partitioned quasi-Newton linear operator, each Bᵢ may use a LBFGS or LSR1 linear operator.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(eplom)

Returns the vector of every elemental element linear operator eplom.eelom_set.

PartitionedStructures.M_abstract_part_struct.get_ee_struct — Method

get_ee_struct(eplom, i)

Returns the i-th elemental element linear operator eplom.eelom_set[i].

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

Build for each index i (∈ {1, ..., n}) a list of the elements using the i-th variable.

PartitionedStructures.M_part_mat.get_eelom_set — Method

get_eelom_set(eplom)

Returns the vector of every elemental element linear operator eplom.eelom_set.

PartitionedStructures.M_part_mat.get_eelom_set — Method

get_eelom_set(eplom :: Elemental_plom_bfgs{T}, i :: Int)

Returns the i-th elemental element linear operator eplom.eelom_set[i].

PartitionedStructures.M_part_mat.set_eelom_set! — Method

set_eelom_set!(eplom, i, eelom)

Sets the i-th elemental elemental linear operator of eplom to eelom.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix of eplom in eplom.spm from the blocs eplom.eelom_set. The sparse matrix is build according to the indices of each elemental element linear operator.

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom — Method

PLBFGSR1_eplom(; n, type, nie, overlapping, prob)

Creates an elemental partitionned limited-memory operator PLSE of N (deduced from n and nie) elemental element linear operators. Each element overlaps the coordinates of the next element by overlapping components. Each element is randomly (rand() > p) choose between an elemental element LBFGS operator or an elemental element LSR1 operator.

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom_rand — Method

PLBFGS_eplom_rand(N,n; type, nie)

Create an elemental partitionned limited-memory operator PLSE of N elemental element linear operators. The size of each element is nie, whose positions are random in the range 1:n. Each element is randomly (rand() > p) choose between an elemental element LBFGS operator or an elemental element LSR1 operator.

PartitionedStructures.ModElemental_plom.get_L — Method

get_L(eplom)

Returns the sparse matrix eplom.L, whose aim to store a cholesky factor. By default eplom.L is not instantiate.

PartitionedStructures.ModElemental_plom.get_L — Method

get_L(eplom, i, j)

Returns the value eplom.L[i,j], from the sparse matrix eplom.L.

PartitionedStructures.ModElemental_plom.get_eelom_set_Bie — Method

get_eelom_set_Bie(eplom, i)

Get the linear operator of the i-th elemental element linear operator of eplom.

PartitionedStructures.ModElemental_plom.get_eelom_sub_set — Method

get_eelom_set_Bie(eplom, indices)

Returns a subset of the elemental element linear operators composing eplom. indices selects the differents elemental element linear operators needed.

PartitionedStructures.ModElemental_plom.identity_eplom_LOSE — Method

identity_eplom_LOSE(vec_indices, N, n; T=T)

Creates an elemental partitionned limited-memory operator of N elemental element linear operators initialized with LBFGS operators. The positions are given by the vector of the element variables element_variables.

PartitionedStructures.ModElemental_plom.set_L! — Method

set_L!(eplom, i, j, value)

Sets the value of eplom.L[i,j] = value.

PartitionedStructures.ModElemental_plom.set_L_to_spm! — Method

set_L_to_spm!(eplom, i, j, value)

Sets the sparse matrix eplom.L to the sparse matrix eplom.spm.

PartitionedStructures.Instances.create_epv_eplom — Method

(eplom,epv) = create_epv_eplom_sr1(;n=9,nie=5,overlapping=1, mul_v=100.)

Create an elemental partitioned limited-memory quasi-Newton operator eplom and elemental partitioned vector epv. Both have the same partitioned structure defined by the size of the problem n, the size of the element nie and the overlapping between the consecutive element overlapping. Each elemental element linear operator is instantiated as a LBFGSOperator, but it may change to a LSR1Operator later on. Each elemental element vector is made from rand(nie) .* mul_v. Warning: You have to choose carefully the values n, nie and overlap otherwise the method may fail.

PartitionedStructures.Instances.create_epv_eplom_bfgs — Method

(eplom,epv) = create_epv_eplom_bfgs(;n=9,nie=5,overlapping=1, mul_v=100.)

Create an elemental partitioned limited-memory quasi-Newton operator PLBFGS eplom and elemental partitioned vector epv. Both have the same partitioned structure defined by the size of the problem n, the size of the element nie and the overlapping between the consecutive element overlapping. Each elemental element linear operator is a LBFGSOperator. Each elemental element vector is made from rand(nie) .* mul_v. Warning: You have to choose carefully the values n, nie and overlap otherwise the method may fail.

PartitionedStructures.Instances.create_epv_eplom_sr1 — Method

(eplom,epv) = create_epv_eplom_sr1(;n=9,nie=5,overlapping=1, mul_v=100.)

Create an elemental partitioned limited-memory quasi-Newton operator PLSR1 eplom and elemental partitioned vector epv. Both have the same partitioned structure defined by the size of the problem n, the size of the element nie and the overlapping between the consecutive element overlapping. Each elemental element linear operator is a LSR1Operator. Each elemental element vector is made from rand(nie) .* mul_v. Warning: You have to choose carefully the values n, nie and overlap otherwise the method may fail.

PartitionedStructures.Instances.create_epv_epm — Method

(epm,epv) = create_epv_epm(;n=9,nie=5,overlapping=1,mul_m=5., mul_v=100.)

Create an elemental partitioned matrix epm and elemental partitioned vector epv. Both have the same partitioned structure defined by the size of the problem n :: Int, the size of the element nie :: Int and the overlapping between the consecutives element overlapping :: Int. Each elemental element matrix is fill with ones, except the terms of the diagonal which are of value mul_v :: Real. The value of each elemental element vector is made from rand(nie) .* mul_v :: Real. Warning: You have to choose carefully the values n, nie and overlap otherwise the method may fail.

PartitionedStructures.Instances.create_epv_epm_rand — Method

(epm,epv) = create_epv_epm_rand(;n=9,nie=5,overlapping=1,range_mul_m=nie:2*nie, mul_v=100.)

Create an partitioned quasi-Newton operator epm and elemental partitioned vector epv. Both have the same partitioned structure defined by the size of the problem n :: Int, the size of the element nie :: Int and the overlapping between the consecutive element overlapping :: Int. Each elemental element matrix is fill with ones, except the terms of the diagonal which are a random value from rand(range_mul_v). Each elemental element vector is made from rand(nie) .* mul_v :: Real. Warning: You have to choose carefully the values n, nie and overlap otherwise the method may fail.

PartitionedStructures.PartMatInterface.update! — Method

update!(epm, epv_y, s)

Update the elemental partitioned operator epm with a partitioned quasi-Newton update considering the difference of elemental partitioned gradients epv_y and the step s. The PSE update is run by default, you can apply a PBFGS or a PSR1 update with the optionnal argument name, respectively name=:pbfgs or name=:psr1.

PartitionedStructures.PartMatInterface.update! — Method

update(epm, epv_y, s)

Update the limited-memory elemental partitioned operator eplom with the partitioned quasi-Newton update PLSR1 considering the difference of elemental partitioned gradients epv_y and the step s.

PartitionedStructures.PartMatInterface.update! — Method

update(epm, epv_y, s)

Update the limited-memory elemental partitioned operator eplom with the partitioned quasi-Newton update PLSR1 considering the difference of elemental partitioned gradients epv_y and the step s.

PartitionedStructures.PartMatInterface.update! — Method

update!(epm, epv_y, s)

Update the limited-memory elemental partitioned operator eplom with the partitioned quasi-Newton update PLSE considering the difference of elemental partitioned gradients epv_y and the step s.

PartitionedStructures.PartMatInterface.update — Method

B = update(epm, epv_y, s)

Update the elemental partitioned operator epm <: Part_mat with a partitioned quasi-Newton update considering the difference of elemental partitioned gradients epv_y and the step s. If epm is an elemental partitioned matrix, the PSE update is run by default. You can apply a PBFGS or a PSR1 update with the optionnal argument name, respectively name=:pbfgs or name=:psr1. It returns a matrix B formed from the updated epm.