Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

PartitionedStructures.M_elt_mat.Counter_elt_mat

PartitionedStructures.M_elt_mat.Elt_mat

PartitionedStructures.M_elt_vec.Elt_vec

PartitionedStructures.M_internal_elt_vec.Internal_elt_vec

PartitionedStructures.M_internal_pv.Internal_pv

PartitionedStructures.M_part_mat.Part_LO_mat

PartitionedStructures.M_part_mat.Part_mat

PartitionedStructures.M_part_v.Part_v

PartitionedStructures.ModElemental_elom_bfgs.Elemental_elom_bfgs

PartitionedStructures.ModElemental_elom_sr1.Elemental_elom_sr1

PartitionedStructures.ModElemental_ev.Elemental_elt_vec

PartitionedStructures.ModElemental_plom.Elemental_plom

PartitionedStructures.ModElemental_plom_bfgs.Elemental_plom_bfgs

PartitionedStructures.ModElemental_plom_sr1.Elemental_plom_sr1

PartitionedStructures.ModElemental_pm.Elemental_pm

PartitionedStructures.ModElemental_pv.Elemental_pv

Base.permute!

PartitionedStructures.Instances.create_epv_eplom

PartitionedStructures.Instances.create_epv_eplom_bfgs

PartitionedStructures.Instances.create_epv_eplom_sr1

PartitionedStructures.Instances.create_epv_epm

PartitionedStructures.Instances.create_epv_epm_rand

PartitionedStructures.Link.eplom_lbfgs_from_epv

PartitionedStructures.Link.eplom_lose_from_epv

PartitionedStructures.Link.eplom_lsr1_from_epv

PartitionedStructures.Link.epm_from_epv

PartitionedStructures.Link.epv_from_epm

PartitionedStructures.Link.mul_epm_epv

PartitionedStructures.Link.mul_epm_epv!

PartitionedStructures.Link.mul_epm_vector

PartitionedStructures.Link.mul_epm_vector!

PartitionedStructures.Link.mul_epm_vector!

PartitionedStructures.Link.string_counters_iter

PartitionedStructures.Link.string_counters_total

PartitionedStructures.M_1_parallel.subproblem

PartitionedStructures.M_2_parallel.subproblem!

PartitionedStructures.M_3_parallel.subproblem3!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_elt_mat.get_Bie

PartitionedStructures.M_elt_mat.update_counter_elt_mat!

PartitionedStructures.M_frontale.frontale!

PartitionedStructures.M_internal_pv.rand_ipv

PartitionedStructures.M_okoubi_koko.okoubi

PartitionedStructures.M_okoubi_koko.okoubi!

PartitionedStructures.M_part_mat.get_spm

PartitionedStructures.M_part_mat.hard_reset_L!

PartitionedStructures.M_part_mat.hard_reset_spm!

PartitionedStructures.M_part_mat.reset_L!

PartitionedStructures.M_part_mat.reset_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_v.add_v!

PartitionedStructures.M_part_v.build_v

PartitionedStructures.M_part_v.build_v!

PartitionedStructures.M_part_v.set_v!

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom_rand

PartitionedStructures.ModElemental_elom_bfgs.init_eelom_LBFGS

PartitionedStructures.ModElemental_elom_bfgs.reset_eelom_bfgs!

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom_rand

PartitionedStructures.ModElemental_elom_sr1.init_eelom_LSR1

PartitionedStructures.ModElemental_elom_sr1.reset_eelom_sr1!

PartitionedStructures.ModElemental_em.create_id_eem

PartitionedStructures.ModElemental_em.fixed_ones_eem

PartitionedStructures.ModElemental_em.identity_eem

PartitionedStructures.ModElemental_em.one_size_bloc

PartitionedStructures.ModElemental_em.ones_eem

PartitionedStructures.ModElemental_ev.create_eev

PartitionedStructures.ModElemental_ev.eev_from_sparse_vec

PartitionedStructures.ModElemental_ev.new_eev

PartitionedStructures.ModElemental_ev.ones_eev

PartitionedStructures.ModElemental_ev.sparse_vec_from_eev

PartitionedStructures.ModElemental_ev.specific_ones_eev

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom_rand

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom_rand

PartitionedStructures.ModElemental_plom_bfgs.identity_eplom_LBFGS

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom_rand

PartitionedStructures.ModElemental_plom_sr1.identity_eplom_LSR1

PartitionedStructures.ModElemental_pm.correlated_var

PartitionedStructures.ModElemental_pm.identity_epm

PartitionedStructures.ModElemental_pm.identity_epm

PartitionedStructures.ModElemental_pm.n_i_SPS

PartitionedStructures.ModElemental_pm.n_i_sep

PartitionedStructures.ModElemental_pm.ones_epm

PartitionedStructures.ModElemental_pm.ones_epm_and_id

PartitionedStructures.ModElemental_pm.part_mat

PartitionedStructures.ModElemental_pv.add_epv!

PartitionedStructures.ModElemental_pv.create_epv

PartitionedStructures.ModElemental_pv.epv_from_epv!

PartitionedStructures.ModElemental_pv.epv_from_v

PartitionedStructures.ModElemental_pv.epv_from_v!

PartitionedStructures.ModElemental_pv.minus_epv!

PartitionedStructures.ModElemental_pv.ones_kchained_epv

PartitionedStructures.ModElemental_pv.rand_epv

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update

PartitionedStructures.PartitionedLOQuasiNewton.PLSE_update

PartitionedStructures.PartitionedLOQuasiNewton.PLSR1_update

PartitionedStructures.PartitionedLOQuasiNewton.Part_update

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update

PartitionedStructures.PartitionedQuasiNewton.PSE_update

PartitionedStructures.PartitionedQuasiNewton.PSR1_update

PartitionedStructures.Utils.BFGS

PartitionedStructures.Utils.SE

PartitionedStructures.Utils.SR1

​

PartitionedStructures.M_internal_elt_vec.Internal_elt_vec — Type

Type that represents internal element vector.

PartitionedStructures.M_internal_pv.Internal_pv — Type

Type that represents internal partitioned vectors.

PartitionedStructures.M_internal_pv.rand_ipv — Method

new_internal_pv(N,n;nᵢ,T)

Define an internal partitionned vector of N elemental nᵢ-sized vector simulating a n-sized T-vector.

PartitionedStructures.M_elt_vec.Elt_vec — Type

Abstract type representing element vectors.

PartitionedStructures.ModElemental_ev.Elemental_elt_vec — Type

Type that reprenents elemental element vector.

PartitionedStructures.ModElemental_ev.create_eev — Method

create_eev(vector_indices)

Create a random elemental element vector from vector_indices.

PartitionedStructures.ModElemental_ev.eev_from_sparse_vec — Method

eev_from_sparse_vec(sparse_vec)

eevfromsparsevec is an interface with SparseArrays.SparseVector. The indices and the values of the elemental element vector are define using findnz(sparsevec).

PartitionedStructures.ModElemental_ev.new_eev — Method

new_eev(nie; T, n)

Create an elemental element vector, with nie randoms values placed at indices within the range 1:n.

PartitionedStructures.ModElemental_ev.ones_eev — Method

ones_eev(nie; T, n)

Create an elemental element vector, with nie values at 1 placed at indices within the range 1:n.

PartitionedStructures.ModElemental_ev.sparse_vec_from_eev — Method

sparse_vec_from_eev(eev)

Create a sparse vector from the element element vector eev.

PartitionedStructures.ModElemental_ev.specific_ones_eev — Method

specific_ones_eev(nie, index; T, mul)

Create an elemental element vector, with nie randoms values multiplied by mul placed at indices in range index:index+nie, with .

PartitionedStructures.ModElemental_pv.Elemental_pv — Type

A type that represent elemental partitioned vector.

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(epm)

initializecomponentlist! Build for each index i (∈ {1,...,n}) the list of the blocs using i.

PartitionedStructures.M_part_v.build_v! — Method

build_v!(pv)

Build from pv the vector v according to the information of each {evᵢ}ᵢ

PartitionedStructures.ModElemental_pv.add_epv! — Method

add_epv!(epv1,epv2)

Build in place of epv2 the addition of epv1 and epv2. Concretely each corresponding elemental vector will be add.

PartitionedStructures.ModElemental_pv.create_epv — Method

create_elemental_pv(elt_ev_set)

create an elemental partitioned vector from a collection (vector) of: sparse vector, elemental element vector or a vector of indices.

PartitionedStructures.ModElemental_pv.epv_from_epv! — Method

epv_from_epv!(evp1, epv2)

Set the elemental partitioned vector epv1 to epv2.

PartitionedStructures.ModElemental_pv.epv_from_v! — Method

epv_from_v!(epv, x)

Set the values of the element partitioned vector epv to x. Usefull to define Uᵢ x, ∀ x.

PartitionedStructures.ModElemental_pv.epv_from_v — Method

epv_from_v(x, epv)

Define a new elemental partitioned vector from x that have the same structure than epv. The value of each elemental element vector comes from the corresponding indices of x. Usefull to define Uᵢ x, ∀ x.

PartitionedStructures.ModElemental_pv.minus_epv! — Method

minus_epv!(epv)

Build in place the -epv, by inversing the value of each elemental element vector.

PartitionedStructures.ModElemental_pv.ones_kchained_epv — Method

ones_kchained_epv(N, k; T)

Construct a N-partitionned k-sized vector such as n = N+k.

PartitionedStructures.ModElemental_pv.rand_epv — Method

new_elemental_pv(N,n;nᵢ,T)

Define an elemental partitionned vector of N elemental nᵢ-sized vector simulating a n-sized T-vector.

PartitionedStructures.M_elt_mat.Counter_elt_mat — Type

Counter_elt_mat

Count for a element matrix the update performed on it, from its definition. total_update + total_reset + total_untouched == iter.

PartitionedStructures.M_elt_mat.Elt_mat — Type

Abstract type representing element matrix

PartitionedStructures.M_elt_mat.get_Bie — Method

get_Bie(elt_mat)

Return the element matrix elt_mat.Bie.

PartitionedStructures.M_elt_mat.update_counter_elt_mat! — Method

update_counter_elt_mat!(cem, qn)

Update the cem counter given the index qn from the quasi-Newton update.

PartitionedStructures.M_part_mat.Part_LO_mat — Type

Abstract type representing partitioned matrix using linear operators

PartitionedStructures.M_part_mat.Part_mat — Type

Abstract type representing partitioned matrix

PartitionedStructures.M_part_mat.get_spm — Method

get_spm(pm)

get_spm(pm, i, j)

Get either the sparse matrix associated to the partitioned matrix pm or ones ot its element at coordinate [i,j].

PartitionedStructures.M_part_mat.hard_reset_L! — Method

hard_reset_L!(pm)

Reset the sparse matrix pm.L.

PartitionedStructures.M_part_mat.hard_reset_spm! — Method

hard_reset_spm!(pm)

Reset the sparse matrix pm.spm.

PartitionedStructures.M_part_mat.reset_L! — Method

reset_L!(pm)

Set the elements of sparse matrix pm.L to 0.

PartitionedStructures.M_part_mat.reset_spm! — Method

reset_spm!(pm)

Set the elements of sparse matrix pm.spm to 0.

PartitionedStructures.M_part_v.Part_v — Type

Abstract type representing partitioned vectors.

PartitionedStructures.M_part_v.add_v! — Method

add_v!(pv, i, value)

add_v!(pv, indices, values)

Add value (resp values) to the vector associated to the partitioned vector pv.v at the indice i (resp indices).

PartitionedStructures.M_part_v.build_v — Method

build_v(pv)

Build the vector v from the partitionned vector pv. Call specialised method depending the type of the element vector inside pv For now if there is mix of elemental and internal element vectors it must be previously transform as internal partitioned vector.

PartitionedStructures.M_part_v.set_v! — Method

set_v!(pv, vec)

set_v!(pv, index, value)

Set to value the vector associated to the partitioned vector pv.v at the indice index. Set to vec the vector associated to the partitioned vector pv.v.

PartitionedStructures.ModElemental_pm.Elemental_pm — Type

Symmetric elemental partitioned matrix

Base.permute! — Method

permute!(epm, p)

Apply the permutation p to the elemental partitionned matrix epm. The permutation is applied to each eem via indices. The current epm permutation is stored in epm.permutation.

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(epm)

Build for each index i (∈ {1, ..., n}) the list of the elemental element matrices using i.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(epm)

Build the sparse matrix epm.spm from the blocs epm.eem_set, according to the elemental element matrices eem.indices.

PartitionedStructures.ModElemental_pm.correlated_var — Method

correlated_var(epm, i)

Get the linked variables to i depending the structure of epm.

PartitionedStructures.ModElemental_pm.identity_epm — Method

identity_epm(N, n; T=T, nie)

Create a partitionned matrix of type T. The partitionned matrix possess N identity elemental element matrices of size nie with randoms positions.

PartitionedStructures.ModElemental_pm.identity_epm — Method

identity_epm(element_variables, N, n; T=T)

Create a partitionned matrix of type T. The partitionned matrix possess N identity elemental element matrices based from the indices of each component of element_variables.

PartitionedStructures.ModElemental_pm.n_i_SPS — Method

n_i_SPS(n)

Define a elemental partitioned matrix of size n. The partitioned matrix is composed by approximatively (n/nie)*2 elemental element matrices. The partitioned matrix is generally not singular.

PartitionedStructures.ModElemental_pm.n_i_sep — Method

n_i_sep(n; T=T, nie=nie, mul=mul)

Define a partitioned nᵢ bloc separable matrix. Each elemental element matrix is composed of 1 except the diagonal that have value mul.

PartitionedStructures.ModElemental_pm.ones_epm — Method

ones_epm(N, n; T=T, nie=nie)

Create a partitionned matrix of type T. The partitionned matrix possess N elemental element matrices ones(nie, nie) whose positions are randoms. The partitionned matrix created may be singular.

PartitionedStructures.ModElemental_pm.ones_epm_and_id — Method

ones_epm_and_id(N, n; T=T, nie=nie)

Create a partitionned matrix of type T. The partitionned matrix possess N+n elemental element matrices. N elemental element matrices ones(nie, nie) whose positions are randoms. The remaining n elemental element matrices are of value [1] , they are all over the diagonal. This way, the partitionned matrix is generally not singular.

PartitionedStructures.ModElemental_pm.part_mat — Method

part_mat(;n=n, T=T, nie=nie, overlapping=overlapping, mul=mul)

Define a elemental partitioned matrix form with nie elemental element matrices. Each elemental element matrix overlap the previous one and the next one by overlapping.

PartitionedStructures.ModElemental_em.create_id_eem — Method

create_id_eem(indices; T=T)

Create an nie identity elemental element matrix of type T at the indices.

PartitionedStructures.ModElemental_em.fixed_ones_eem — Method

fixed_ones_eem(index, nie; type=T, mul=mul)

Create an nie elemental element matrix of type T at indices index:index+nie-1. All element have the value 1 except the diagonal that have the value mul, it is use to define diagonal dominant matrix.

PartitionedStructures.ModElemental_em.identity_eem — Method

identity_eem(nie; T=T, n=n)

Create an nie identity elemental element matrix of type T at random indices in the range 1:n.

PartitionedStructures.ModElemental_em.one_size_bloc — Method

one_size_bloc(i)

Define a elemental element matrix of type T of size one at the index i.

PartitionedStructures.ModElemental_em.ones_eem — Method

ones_eem(nie; T=T, n=n)

Create an nie ones elemental element matrix of type T at random indices in the range 1:n.

PartitionedStructures.M_okoubi_koko.okoubi! — Method

okoubi(epm, epv, epv_tmp, res)

Solve the partitioned linear problem using the Okoubi and Koko method described in the paper available https://doi.org/10.1007/s11590-016-1020-x. This method avoid the allocation of epv_tmp and res.

PartitionedStructures.M_okoubi_koko.okoubi — Method

okoubi(epm, epv)

Solve the partitioned linear problem using the Okoubi and Koko method described in the paper available https://doi.org/10.1007/s11590-016-1020-x.

PartitionedStructures.M_frontale.frontale! — Method

frontale!(epm)

Produce the Cholesky frontal factorization of the elemental partitioned matrix epm. The sparse factor L is stored in epm.L.

PartitionedStructures.M_1_parallel.subproblem — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.M_2_parallel.subproblem! — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.M_3_parallel.subproblem3! — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.Utils.BFGS — Method

BFGS(s, y, B)

Perform the BFGS update over the matrix B by using the vector s and y.

PartitionedStructures.Utils.SE — Method

SE(s, y, B)

Perform either the BFGS update or SR1 over the matrix B by using the vector s and y.

PartitionedStructures.Utils.SR1 — Method

SR1(s, y, B)

Perform the BFGS update over the matrix B by using the vector s and y.

PartitionedStructures.Link.eplom_lbfgs_from_epv — Method

eplom_lbfgs_from_epv(epm)

Create an elemental partitioned linear operator matrix with the same partitioned structure than epv. Each elemental element linear operator is set with an LBFGS operator.

PartitionedStructures.Link.eplom_lose_from_epv — Method

eplom_lose_from_epv(epm)

Create an elemental partitioned linear operator matrix with the same partitioned structure than epv. Each elemental element linear operator is set with an LSR1 operator.

PartitionedStructures.Link.eplom_lsr1_from_epv — Method

eplom_lsr1_from_epv(epm)

Create an elemental partitioned linear operator matrix with the same partitioned structure than epv. Each elemental element linear operator is set with an LSR1 operator.

PartitionedStructures.Link.epm_from_epv — Method

epm_from_epv(epm)

Create an elemental partitioned matrix with the same partitioned structure than epv. Each elemental element matrix is set with an identity matrix.

PartitionedStructures.Link.epv_from_epm — Method

epv_from_epm(epm)

Create an elemental partitioned vector with the same partitioned structure than epm.

PartitionedStructures.Link.mul_epm_epv! — Method

mul_epm_epv!(epv_res, epm, epv)

Compute the product between the elemental partitioned matrix epm and the elemental partitioned vecto epv. The result is stored in the elemental partitioned vector epv_res.

PartitionedStructures.Link.mul_epm_epv — Method

mul_epm_epv(epm, epv)

Compute the product between the elemental partitioned matrix epm and the elemental partitioned vecto epv.

PartitionedStructures.Link.mul_epm_vector! — Method

mul_epm_vector!(res, epm, epv, x)

Compute the product between the elemental partitioned matrix epm and the vector x. The result is stored in the vector res.

PartitionedStructures.Link.mul_epm_vector! — Method

mul_epm_vector!(res, epm, x)

Compute the product between the elemental partitioned matrix epm and the vector x. The result is stored in the vector res.

PartitionedStructures.Link.mul_epm_vector — Method

mul_epm_vector(epm, x)

Compute the product between the elemental partitioned matrix epm and the vector x.

PartitionedStructures.Link.string_counters_iter — Method

string_counters_iter(pm)

Produce a string that summarize the partitioned update applied onto pm at the last iterate.

PartitionedStructures.Link.string_counters_total — Method

string_counters_total(pm)

Produce a string that summarize the partitioned update applied onto pm from its allocation.

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update — Method

PBFGS_update(epm_B, s, epv_y)

Define the partitioned BFGS update of the partioned matrix epmB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedQuasiNewton.PSE_update — Method

PSE_update(epm_B, s, epv_y)

Define the partitioned SR1 update of the elemental partioned matrix epmB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedQuasiNewton.PSR1_update — Method

PSR1_update(epm_B, s, epv_y)

Define the partitioned SR1 update of the elemental partioned matrix epmB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update — Method

PLBFGS_update(eplom_B, s, epv_y)

Define the partitioned LBFGS update of the partioned matrix eplomB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedLOQuasiNewton.PLSE_update — Method

PLSE_update(eplom_B, epv_y, s)

Perform the partitionned update of eplomB. eplomB is build from LBFGS or LSR1 elemental element matrices. The update performed on eachh element matrix correspond to the linear operator associated.

PartitionedStructures.PartitionedLOQuasiNewton.PLSR1_update — Method

PLSR1_update(eplom_B, s, epv_y)

Define the partitioned LSR1 update of the partioned matrix eplomB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedLOQuasiNewton.Part_update — Method

Part_update(eplom_B, epv_y, s)

Perform the partitionned update of eplomB. eplomB is build from LBFGS or LSR1 elemental element matrices. The update performed on eachh element matrix correspond to the linear operator associated.

PartitionedStructures.ModElemental_elom_bfgs.Elemental_elom_bfgs — Type

Type that represents an elemental element linear operator LBFGS

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom — Method

LBFGS_eelom(nie, T=T, index=index)

Create a Elemental_elom_bfgs of type T of size nie, the indices are in the range index:index+nie-1.

PartitionedStructures.ModElemental_elom_bfgs.LBFGS_eelom_rand — Method

LBFGS_eelom_rand(nie, T=T, n=n)

Create a Elemental_elom_bfgs of type T with nie random indices within the range 1:n.

PartitionedStructures.ModElemental_elom_bfgs.init_eelom_LBFGS — Method

init_eelom_LBFGS(indices; T=T)

Define a Elemental_elom_bfgs of type Elemental_elom_sr1 based from the vector indices.

PartitionedStructures.ModElemental_elom_bfgs.reset_eelom_bfgs! — Method

reset_eelom_bfgs!(eelom)

Reset the LBFGS linear operator of the elemental element linear operator matrix.

PartitionedStructures.ModElemental_plom_bfgs.Elemental_plom_bfgs — Type

Type that represents an element partitioned matrix defined with elemental element linear operators LBFGS

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

Build for each index i (∈ {1, ..., n}) the list of the blocs using i.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix eplom.spm from the blocs eplom.eelom_set, according to the indices of each elemental element linear operator matrice.

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom — Method

PLBFGS_eplom(N, n; type, nie)

Create an elemental partitionned limited memory of N elemental element linear operators matrices which are overlapping the next block coordinates by overlapping.

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom_rand — Method

PLBFGS_eplom_rand(N, n; type, nie)

Create an elemental partitionned limited memory matrix of N elemental element linear operators matrices whose positions are random.

PartitionedStructures.ModElemental_plom_bfgs.identity_eplom_LBFGS — Method

identity_eplom_LBFGS(vec_indices, N, n; T=T)

Create an elemental partitionned limited memory of N elemental element linear operators matrices whose the positions are given by vec_indices.

PartitionedStructures.ModElemental_elom_sr1.Elemental_elom_sr1 — Type

Type that represents elemental element linear operator LSR1

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom — Method

LSR1_eelom(nie, T=T, index=index)

Create a Elemental_elom_sr1 of type T of size nie, the indices are in the range index:index+nie-1.

PartitionedStructures.ModElemental_elom_sr1.LSR1_eelom_rand — Method

LSR1_eelom_rand(nie, T=T, n=n)

Create a Elemental_elom_sr1 of type T with nie random indices within the range 1:n.

PartitionedStructures.ModElemental_elom_sr1.init_eelom_LSR1 — Method

init_eelom_LSR1(indices; T=T)

Define a Elemental_elom_sr1 of type Elemental_elom_sr1 based from the vector indices.

PartitionedStructures.ModElemental_elom_sr1.reset_eelom_sr1! — Method

index_eelom_sr1!(eelom)

Reset the LSR1 linear operator of the elemental element linear operator matrix.

PartitionedStructures.ModElemental_plom_sr1.Elemental_plom_sr1 — Type

Type that represents an element partitioned matrix defined with elemental element linear operators LBFGS

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

Build for each index i (∈ {1, ..., n}) the list of the blocs using i.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix eplom.spm from the blocs eplom.eelom_set, according to the indices of each elemental element linear operator matrice.

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom — Method

PLSR1_eplom(N, n; type, nie)

Create an elemental partitionned limited memory of N elemental element linear operators matrices which are overlapping the next block coordinates by overlapping.

PartitionedStructures.ModElemental_plom_sr1.PLSR1_eplom_rand — Method

PLSR1_eplom_rand(N, n; type, nie)

Create an elemental partitionned limited memory matrix of N elemental element linear operators matrices whose positions are random.

PartitionedStructures.ModElemental_plom_sr1.identity_eplom_LSR1 — Method

identity_eplom_LSR1(vec_indices, N, n; T=T)

Create an elemental partitionned limited memory of N elemental element linear operators matrices whose the positions are given by vec_indices.

PartitionedStructures.ModElemental_plom.Elemental_plom — Type

Type that represents an element partitioned matrix defined with elemental element linear operators LBFGS/LSR1

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

initializecomponentlist! Build for each index i (∈ {1,...,n}) the list of the blocs using i.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix spm from the blocs eplom.eelom_set, according to the indices.

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom — Method

PLBFGS_eplom(N,n; type, nie) Create a a partitionned limited memory matrix of N LBFGSLinearOperators blocks whose overlap next block coordinates by overlapping.

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom_rand — Method

PLBFGSeplomrand(N,n; type, nie) Create a a partitionned limited memory matrix of N LBFGSLinearOperators blocs whose positions are randoms

PartitionedStructures.Instances.create_epv_eplom — Method

create_epv_epm_rand(;n=n,nie=nie,overlpapping=overlapping, mul_m=mul_m, mul_v=mul_v)

Create a elemental partitioned vector and a random elemental partitioned matrix using linear operators LBFGS/LSR1 with the same partitioned structure defined by n,nie,overlapping,mul_l,mul_v.

PartitionedStructures.Instances.create_epv_eplom_bfgs — Method

create_epv_eplom_bfgs(;n=n,nie=nie,overlpapping=overlapping, mul_m=mul_m, mul_v=mul_v)

Create a elemental partitioned vector and a random elemental partitioned matrix using linear operators LBFGS with the same partitioned structure defined by n,nie,overlapping,mul_l,mul_v.

PartitionedStructures.Instances.create_epv_eplom_sr1 — Method

create_epv_eplom_bfgs(;n=n,nie=nie,overlpapping=overlapping, mul_m=mul_m, mul_v=mul_v)

Create a elemental partitioned vector and a random elemental partitioned matrix using linear operators LBFGS with the same partitioned structure defined by n,nie,overlapping,mul_l,mul_v.

PartitionedStructures.Instances.create_epv_epm — Method

create_epv_epm(;n=n,nie=nie,overlpapping=overlapping, mul_m=mul_m, mul_v=mul_v)

Create an elemental partitioned vector and a elemental partitioned matrix with the same partitioned structure defined by n,nie,overlapping,mul_l,mul_v.

PartitionedStructures.Instances.create_epv_epm_rand — Method

create_epv_epm_rand(;n=n,nie=nie,overlpapping=overlapping, mul_m=mul_m, mul_v=mul_v)

Create a random elemental partitioned vector and a random elemental partitioned matrix with the same partitioned structure defined by n,nie,overlapping,mul_l,mul_v.