Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

Base.permute!

PartitionedStructures.M_1_parallel.subproblem

PartitionedStructures.M_2_parallel.subproblem!

PartitionedStructures.M_3_parallel.subproblem3!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_abstract_part_struct.initialize_component_list!

PartitionedStructures.M_internal_pv.rand_ipv

PartitionedStructures.M_part_mat.reset_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_mat.set_spm!

PartitionedStructures.M_part_v.build_v

PartitionedStructures.M_part_v.build_v!

PartitionedStructures.ModElemental_ev.eev_from_sparse_vec

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom_rand

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom_rand

PartitionedStructures.ModElemental_pm.correlated_var

PartitionedStructures.ModElemental_pm.identity_epm

PartitionedStructures.ModElemental_pm.n_i_SPS

PartitionedStructures.ModElemental_pm.n_i_sep

PartitionedStructures.ModElemental_pm.ones_epm

PartitionedStructures.ModElemental_pm.ones_epm_and_id

PartitionedStructures.ModElemental_pm.part_mat

PartitionedStructures.ModElemental_pv.add_epv!

PartitionedStructures.ModElemental_pv.create_epv

PartitionedStructures.ModElemental_pv.minus_epv!

PartitionedStructures.ModElemental_pv.ones_kchained_epv

PartitionedStructures.ModElemental_pv.rand_epv

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update

PartitionedStructures.PartitionedLOQuasiNewton.Part_update

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update

PartitionedStructures.PartitionedQuasiNewton.PSR1_update

PartitionedStructures.Utils.BFGS

PartitionedStructures.Utils.SR1

​

PartitionedStructures.M_internal_pv.rand_ipv — Method

new_internal_pv(N,n;nᵢ,T)

Define an internal partitionned vector of N elemental nᵢ-sized vector simulating a n-sized T-vector.

PartitionedStructures.ModElemental_ev.eev_from_sparse_vec — Method

SparseArrays.SparseVector interface

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(epm)

initializecomponentlist! Build for each index i (∈ {1,...,n}) the list of the blocs using i.

PartitionedStructures.M_part_v.build_v! — Method

build_v!(pv)

Build from pv the vector v according to the information of each {evᵢ}ᵢ

PartitionedStructures.ModElemental_pv.add_epv! — Method

add_epv!(epv1,epv2)

Build in place of epv2 the addition of epv1 and epv2. Concretely each corresponding elemental vector will be add.

PartitionedStructures.ModElemental_pv.create_epv — Method

create_elemental_pv(elt_ev_set)

create easily a pv from eltevset (confort)

PartitionedStructures.ModElemental_pv.minus_epv! — Method

minus_epv!(epv)

Build in place the -epv, by inversing the value of each elemental element vector.

PartitionedStructures.ModElemental_pv.ones_kchained_epv — Method

ones_kchained_epv(N, k; T)

Construct a N-partitionned k-sized vector such as n = N+k. eltype(pv)=T

PartitionedStructures.ModElemental_pv.rand_epv — Method

new_elemental_pv(N,n;nᵢ,T)

Define an elemental partitionned vector of N elemental nᵢ-sized vector simulating a n-sized T-vector.

PartitionedStructures.M_part_mat.reset_spm! — Method

reset_spm!(pm) Reset the sparse matrix pm.spm

PartitionedStructures.M_part_v.build_v — Method

build_v(pv)

Build the vector v from the partitionned vector pv. Call specialised method depending the type of the element vector inside pv For now if there is mix of elemental and internal element vectors it must be previously transform as internal partitioned vector.

Base.permute! — Method

permute!(epm,p)

apply the permutation p to the elemental partitionned matrix epm. The permutation is applied to each eem via indices. The current epm permutation is stored in epm.permutation

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(epm)

initializecomponentlist! Build for each index i (∈ {1,...,n}) the list of the blocs using i.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(epm)

Build the sparse matrix spm from the blocs epm.eem_set, according to the indinces.

PartitionedStructures.ModElemental_pm.correlated_var — Method

correlated_var(epm,i)

correlated_var(epm,i) get the linked vars to i depending the structure of epm.

PartitionedStructures.ModElemental_pm.identity_epm — Method

identity_epm(N,n; type, nie)

Create a a partitionned matrix of N nie-identity blocs whose positions are randoms

PartitionedStructures.ModElemental_pm.n_i_SPS — Method

n_i_SPS(n)

A nᵢ bloc separable matrix By default nᵢ = 5

PartitionedStructures.ModElemental_pm.n_i_sep — Method

n_i_diag_dom(n)

A nᵢ bloc separable matrix By default nᵢ = 5

PartitionedStructures.ModElemental_pm.ones_epm — Method

ones_epm(N,n; type, nie)

Create a a partitionned matrix of N ones(nie,nie) blocs whose positions are randoms. Be careful the partitionned matrix created may be singular.

PartitionedStructures.ModElemental_pm.ones_epm_and_id — Method

ones_epm_and_id(N,n; type, nie)

Create a a partitionned matrix of N ones(nie,nie) blocs whose positions are randoms and adding . Not singular.

PartitionedStructures.ModElemental_pm.part_mat — Method

part_mat(n)

A nᵢ partially bloc separable matrix, whith on the diagonal band regular with overlapping (=1 by default) By default nᵢ=5, overlapping=1, mul=5.

PartitionedStructures.M_1_parallel.subproblem — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.M_2_parallel.subproblem! — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.M_3_parallel.subproblem3! — Method

subproblem(epm_A, epv_b, epv_x, comp_list, i)

define the subproblem which must be solve for the i-th variable

PartitionedStructures.Utils.BFGS — Method

BFGS(s, y, B)

Perform the BFGS update over the matrix B by using the vector s and y.

PartitionedStructures.Utils.SR1 — Method

SR1(s, y, B)

Perform the BFGS update over the matrix B by using the vector s and y.

PartitionedStructures.PartitionedQuasiNewton.PBFGS_update — Method

PBFGS_update(epm_B, s, epv_y)

Define the partitioned BFGS update of the partioned matrix epmB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedQuasiNewton.PSR1_update — Method

PSR1_update(epm_B, s, epv_y)

Define the partitioned BFGS update of the partioned matrix epmB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedLOQuasiNewton.PLBFGS_update — Method

PLBFGS_update(eplom_B, s, epv_y)

Define the partitioned BFGS update of the partioned matrix eplomB, given the step s and the element gradient difference epvy

PartitionedStructures.PartitionedLOQuasiNewton.Part_update — Method

Part_update(eplom_B, epv_y, s)

Perform the partitionned update of eplomB. eplomB is build from LBFGS or LSR1 elemental element matrices. The update performed on eachh element matrix correspond to the linear operator associated.

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

initializecomponentlist! Build for each index i (∈ {1,...,n}) the list of the blocs using i.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix spm from the blocs eplom.eelom_set, according to the indices.

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom — Method

PLBFGS_eplom(N,n; type, nie)

Create a a partitionned limited memory matrix of N LBFGSLinearOperators blocks whose overlap next block coordinates by overlapping.

PartitionedStructures.ModElemental_plom_bfgs.PLBFGS_eplom_rand — Method

PLBFGS_eplom_rand(N,n; type, nie)

Create a a partitionned limited memory matrix of N LBFGSLinearOperators blocs whose positions are randoms

PartitionedStructures.M_abstract_part_struct.initialize_component_list! — Method

initialize_component_list!(eplom)

initializecomponentlist! Build for each index i (∈ {1,...,n}) the list of the blocs using i.

PartitionedStructures.M_part_mat.set_spm! — Method

set_spm!(eplom)

Build the sparse matrix spm from the blocs eplom.eelom_set, according to the indices.

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom — Method

PLBFGS_eplom(N,n; type, nie)

Create a a partitionned limited memory matrix of N LBFGSLinearOperators blocks whose overlap next block coordinates by overlapping.

PartitionedStructures.ModElemental_plom.PLBFGSR1_eplom_rand — Method

PLBFGS_eplom_rand(N,n; type, nie)

Create a a partitionned limited memory matrix of N LBFGSLinearOperators blocs whose positions are randoms